#88 - refactor : 스트리머 네비게이션 그래프 관련 코드 네비게이션 패키지로 이동, Route 상수 변경
일관성있게 navigation 패키지를 만들어서 작업하였습니다.
그리고 StreamerRoute를 Object에서 상수로 변경해서 사용중입니다.
