+++ compose/streamer/src/main/java/com/hegunhee/compose/streamer/navigation/StreamerNavGraph.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.compose.streamer
+package com.hegunhee.compose.streamer.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.CompositionLocalProvider
@@ -6,21 +6,20 @@ import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.hegunhee.compose.streamer.StreamerScreenRoot
 
 val LocalPaddingValues = compositionLocalOf { PaddingValues(0.dp) }
+
+const val StreamerRoute = "streamer"
 
 fun NavGraphBuilder.streamerNavGraph(
     paddingValues : PaddingValues,
     onNavigateTwitchChannelClick : (String) -> Unit
 ) {
-    composable(route = StreamerNavGraph.streamerRoute) {
+    composable(route = StreamerRoute) {
         CompositionLocalProvider(LocalPaddingValues provides paddingValues) {
             StreamerScreenRoot(onNavigateTwitchChannelClick = onNavigateTwitchChannelClick)
         }
 
     }
-}
-
-object StreamerNavGraph{
-    const val streamerRoute = "streamer"
 }
